back-end/requirements.txt
Django>=5.2.11
psycopg2-binary>=2.9.9
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.0
